Refactor the hiloapi
We should avoid exec at all cost as it's not safe. setattr will do a
fine job in this specific use case.

Wrapping requests under the same method will make it more reliable, with
retries and centralized exception catching.

Unless we have a specific usecase, we shouldn't have a class in a class.

Removing repetitive code.

Also, we need to turn the lights on before setting intensity otherwise
the intensity call is not taken into account.

fixes #25