Async implementation
Based on HA doc [1], we should use the async core.

Also, we're introducing a breaking change. Normally, light entities
should be added as such, and not as switches. We can keep the original
behavior by enabling the config boolean "light_as_switch" but users are
encouraged to migrate their automation and scripts using lights instead
of switches.

[1] https://developers.home-assistant.io/docs/asyncio_working_with_async/