Automatically generating energy meters and adding Gateway device
This commit aims to automatically generate and add the energy meters to
the Energy dashboard. All entities, meters and sensors are automatically
configured.

The details have been described in the README.md.

Also, I'm creating a new gateway device to monitor the online status of
the Hilo gateway.

Finally, I'm also proposing to add a license just to protect the developpers
in case something goes wrong for someone. This project has to be fun and is
not commercially backed so there's no reason we'd take the risk of
exposing ourselves to law suits.

Closes #41, closes #38